🗃️Add sql unit tests + fix types

+++ src/main/scala/com/lion/rafiki/sql/Pagination.scala
@@ -4,10 +4,10 @@ import doobie._
 import doobie.implicits._
 
 trait SQLPagination {
-  def limit[A: Read](lim: Int)(q: Query0[A]): Query0[A] =
+  def limit[A: Read](lim: Long)(q: Query0[A]): Query0[A] =
     (q.toFragment ++ fr" LIMIT $lim").query
 
-  def paginate[A: Read](lim: Int, offset: Int)(q: Query0[A]): Query0[A] =
+  def paginate[A: Read](lim: Long, offset: Long)(q: Query0[A]): Query0[A] =
     (q.toFragment ++ fr" LIMIT $lim OFFSET $offset").query
 }
